refactored store to be dir with files for each context
