+++ client/src/Store/UserContext.tsx
@@ -1,0 +1,10 @@
+export interface UserProps {
+  username: string;
+  name?: string;
+  age?: number;
+  bio?: string;
+  highlightedFavorite?: string;
+  favorites?: string[];
+  follows?: UserProps[];
+  following?: UserProps[];
+}
